test(title-parsing): add unit tests for cleanCifraClubTitle and extractTitleAndArtist functions
